a query3 ainda da alguns erros mas deixei ai por enquanto

+++ trabalho-pratico/include/entidades/user.h
@@ -43,7 +43,7 @@ char* getUserCountry(User* u);
 
 SubscriptionType getUserSubscriptionType(User* u);
 
-int* getUserLikedMusics(User* u);
+long int* getUserLikedMusics(User* u);
 
 int getUserNumLikedMusics(User* u);
